feat: add folder check

+++ index.js
@@ -6,16 +6,23 @@ const chalk = require('chalk');
 const figlet = require('figlet');
 const ora = require('ora');
 
-const projectMap = require('./lib/config');
+const config = require('./lib/config');
 const version = require('./package.json').version;
 
 const run = async () => {
   print();
-  const answers = await askQuestions();
-  create(answers);
+  const answer = await askQuestions();
+  create(getTemplateKey(answer));
 };
 
-const print = () =>
+const getTemplateKey = (answer) => {
+  return {
+    ...answer,
+    template: answer.template.split(' ')[0]
+  }
+}
+
+const print = () => {
   console.log(
     chalk.green(
       figlet.textSync('MUGUA-CLI', {
@@ -25,8 +32,22 @@ const print = () =>
       })
     )
   );
+  console.log(
+    chalk.green(
+      'you can see more templates at: https://github.com/lq782655835/ts-templates'
+    )
+  );
+};
 
 const askQuestions = () => {
+  const maxCharLength = Math.max(
+    ...Object.keys(config).map((key) => key.length)
+  );
+  const strPadEndLength = maxCharLength + 10;
+  const choices = Object.keys(config).map(
+    (key) => `${key.padEnd(strPadEndLength)} ${config[key].description}`
+  );
+
   const prompt = inquirer.createPromptModule();
   const questions = [
     {
@@ -38,20 +59,23 @@ const askQuestions = () => {
       type: 'list',
       name: 'template',
       message: 'choose your project template',
-      choices: Object.keys(projectMap),
+      choices,
     },
   ];
   return prompt(questions);
 };
 
 const create = ({ project, template }) => {
+  const dest = `${shell.pwd()}/${project}`;
+  if (shell.test('-e', dest)) {
+    console.log(chalk.red('目标文件夹已存在'));
+    process.exit(1);
+  }
+
   const spinner = ora('downloading template');
   spinner.start();
 
-  const pwd = shell.pwd();
-  const dest = `${pwd}/${project}`;
-  const { url, description } = projectMap[template];
-  console.log(chalk.yellow(` ${description}`));
+  const { url } = config[template];
   clone(url, dest, null, function () {
     spinner.stop();
     shell.rm('-rf', dest + '/.git');
